requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
websocket-client>=1.6.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
fake-useragent>=1.4.0
pillow>=10.2.0
qrcode>=7.4.0